% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cluster.R
\name{cluster_paths}
\alias{cluster_paths}
\title{Cluster t-T paths}
\usage{
cluster_paths(x, ncluster, dist = c("Hausdorff", "Frechet", "Euclidean"))
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Output of \code{\link[=read_hefty_xlsx]{read_hefty_xlsx()}}.}

\item{ncluster}{an integer scalar or vector with the desired number of groups}

\item{dist}{character; algorithm to calculate a dissimilarity matrix
(distance) for lines; one of \code{Euclidean}, \code{Hausdorff} (the default) or
\code{Frechet}.}
}
\value{
a tibble with the path \code{segment} (integer) and \code{cluster} (factor)
}
\description{
Hierarchical clustering of t-T paths based on the Hausdorff distance
between paths.
}
\examples{
\dontrun{
data(tT_paths)
cluster_paths(tT_paths, ncluster = 3)
}
}
