% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_paths.R
\name{densify_cluster}
\alias{densify_cluster}
\title{Densify clustered paths}
\usage{
densify_cluster(x)
}
\arguments{
\item{x}{clustered t-T paths. Output of \code{\link[=cluster_paths]{cluster_paths()}}.}
}
\value{
tibble
}
\description{
Densify clustered paths
}
\examples{
\dontrun{
data(tT_paths)
tT_paths_subset <- subset(tT_paths, Comp_GOF >= 0.5)
cluster_paths(tT_paths_subset, cluster = 3) |>
merge(tT_paths_subset, by = 'segment') |>
dplyr::group_by(cluster) |>
densify_cluster()
}
}
