% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_paths.R
\name{densify_paths}
\alias{densify_paths}
\title{Densify Paths}
\usage{
densify_paths(
  x,
  GOF_rank = 10L,
  n = 10L,
  max_distance = 1,
  samples = 100L,
  replace = FALSE
)
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Output of \code{\link[=read_hefty_xlsx]{read_hefty_xlsx()}}.}

\item{GOF_rank}{numeric. Selects only the \code{GOF_rank}-th highest GOF ranked paths.
If all GOFs should be used, set to \code{Inf}. Default is \code{10}.}

\item{n}{integer. Adds \code{n} (10 by
default) equally-spaced extra points along each path
segment (between vertices).}

\item{max_distance}{numeric. Adds points at a maximum distance of
\code{max_distance} (in Myr) from each other. 1 by default.}

\item{samples}{integer. Number of random samples of the data. This number should
be less or equal then the \code{nrow(x)}. Be aware that a large number will require
a long(!) processing time. The default is \code{100}.}

\item{replace}{logical. Should sampling be with replacement?}
}
\value{
data.frame
}
\description{
Adds extra points between the vertices of the path segments
}
\examples{
data(tT_paths)
densify_paths(tT_paths)
}
