% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeFTy.SmoothR-package.R
\docType{package}
\name{HeFTy.SmoothR-package}
\alias{HeFTy.SmoothR}
\alias{HeFTy.SmoothR-package}
\title{HeFTy.SmoothR: Analyzing Inverse Thermal History Models}
\description{
Density estimates for HeFTy-modeled time-temperature cooling paths.
}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

Authors:
\itemize{
  \item Joel Padgett \email{joel.padgett@ucalgary.ca} [copyright holder]
}

}
\keyword{internal}
