% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_statistics.R
\name{path_statistics}
\alias{path_statistics}
\title{Path statistics}
\usage{
path_statistics(x)
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Output of \code{\link[=read_hefty]{read_hefty()}}.}
}
\value{
tibble.
}
\description{
Provides binned statistics (mean, median, IQR, quantiles, etc.) of modeled t-T paths.
}
\examples{
data(tT_paths)
path_statistics(tT_paths)
}
