% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_density.R
\name{plot_path_density}
\alias{plot_path_density}
\title{Path Density Plot}
\usage{
plot_path_density(x, bins = 50L, densify = TRUE, ...)
}
\arguments{
\item{x}{data.frame containing the t-T paths. Output of \code{\link[=read_hefty_xlsx]{read_hefty_xlsx()}} or \code{\link[=densify_paths]{densify_paths()}}}

\item{bins}{integer. Amount of bins used for the kernel density estimate. 50 by default.}

\item{densify}{logical. Whether the paths in \code{x} should be densified first?
Default is \code{TRUE}.}

\item{...}{optional arguments passed to \code{\link[=densify_paths]{densify_paths()}} (only if \code{densify==TRUE}).}
}
\value{
ggplot
}
\description{
Creates a 2d kernel density estimate of the t-T paths and plots it using
ggplot
}
\examples{
data(s14MM_v1)
plot_path_density(s14MM_v1)
}
