% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s14MM_v1}
\alias{s14MM_v1}
\title{HeFTy paths example}
\format{
An object of class \code{tibble}
}
\usage{
data('s14MM_v1')
}
\description{
some description
}
\examples{
data("s14MM_v1")
head(s14MM_v1)
}
\keyword{datasets}
