% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_paths.R
\name{densify_paths}
\alias{densify_paths}
\title{Densify Paths}
\usage{
densify_paths(x, GOF_rank = 10L, n = 10L, samples = 100L, replace = FALSE)
}
\arguments{
\item{x}{t-T and GOF data of the modeled paths. Out put of \code{\link[=read_hefty_xlsx]{read_hefty_xlsx()}}.}

\item{GOF_rank}{numeric. Selects only the n highest GOF ranked paths.}

\item{n}{integer. Adds \code{n} equally-spaced extra points along each path
segment (between vertices). The default is 10.}

\item{samples}{integer. Number of random samples of the data. This number should
be less or equal then the\code{nrow(x)}. Be aware that a large number will require
a long(!) processing time. The default is \code{100}.}

\item{replace}{logical. Should sampling be with replacement?}
}
\value{
data.frame
}
\description{
Adds extra points between the vertices of the path segments
}
\examples{
data(s14MM_v1)
densify_paths(s14MM_v1)
}
